pyproject.toml
[tool.poetry]
name = "fishbone-backend"
version = "0.1.0"
description = ""
authors = ["Joshua Godwin <119361939+joshuagodwin@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
google-generativeai = "^0.5.4"
uvicorn = "^0.30.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"